IOT-power/requirements.txt
flask
numpy
firebase_admin
pyserial
apiai